requirements.txt
fastapi==0.101.1
mysql-connector-python==8.1.0
uvicorn==0.23.2
